fixed some boxes not disappearing
bug fix: some boxes that were in front of the lightbulb weren't disappearing or being detected. i tried a couple of different solutions but ultimately it seems like raising the height of the area3D which detects the boxes was all that was needed (hopefully).

I did also redo the collision shape for the boxes manually.